refactor(core): extract error/crypto/encoding/addr types out of lib.rs (#198)
Per epic #182, lib.rs is meant to define the dependency direction and
re-export the crate's API surface — not host cross-cutting definitions.
Move the ~170 LoC of definitions that had accumulated in lib.rs into
dedicated, cohesive modules and keep lib.rs as thin re-exports.

What moved where:
- error.rs    — CoreError, ErrorCode (+ core_error_display test)
- crypto.rs   — secret_key_sign, public_key_verify, generate_secret_key
- encoding.rs — base32_encode, base32_decode, parse_node_id (+ base32/
                node-id tests)
- addr.rs     — node_ticket, parse_node_addr, ParsedNodeAddr

Zero public-API change: lib.rs re-exports every previously-public item at
its same crate path via `pub use`, so `iroh_http_core::CoreError`,
`generate_secret_key`, `base32_encode`, `node_ticket`, etc. resolve
exactly as before. `crate::base32_decode` (pub(crate)) is likewise
preserved for internal callers. Verified with `cargo build --workspace`.

Update tests/architecture.rs canonical-layout allowlist to include the
new utility modules, as its doc comment requires when adding top-level
modules.

Closes #198
